backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0

# AI & LLM
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-groq==0.0.1
langgraph==0.0.20
langsmith==0.0.77
google-generativeai==0.3.2
groq==0.4.1

# Voice
elevenlabs==0.2.26
openai-whisper==20231117
speechrecognition==3.10.1
pydub==0.25.1

# Vector DB & Memory
chromadb==0.4.22
qdrant-client==1.7.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
